[sc-490] New Upg Of Criterion Factory & Consequences
