+++ modules/php/Game/Card/Consequence/Category/Generic/CardUsedConsequence.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SmileLife\Card\Consequence;
+namespace SmileLife\Card\Consequence\Category\Generic;
 
 /**
  * Description of CardUsedConsequence
