+++ modules/php/Game/Card/Consequence/Category/Love/FlirtOnAdulteryConsequence.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SmileLife\Card\Consequence;
+namespace SmileLife\Card\Consequence\Category\Love;
 
 use Core\Models\Player;
 use SmileLife\Card\CardManager;
@@ -12,7 +12,7 @@ use SmileLife\Card\Consequence\ConsequenceException;
  *
  * @author Mr_Kywar mr_kywar@gmail.com
  */
-class FlirtDoublonDectectionConcequence extends Consequence {
+class FlirtOnAdulteryConsequence extends Consequence {
 
     /**
      * 
@@ -39,7 +39,7 @@ class FlirtDoublonDectectionConcequence extends Consequence {
     }
 
     public function execute() {
-        throw new ConsequenceException("Consequence-FDDC : Not Yet implemented");
+        throw new ConsequenceException("Consequence-FOAC : Not Yet implemented");
     }
 
 }
