+++ modules/php/Game/Card/Consequence/Category/Studies/LimitlessStudieConsequence.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace SmileLife\Card\Consequence\Category\Studies;
+
+use SmileLife\Card\Consequence\ConsequenceException;
+
+/**
+ * Description of LimitlessStudieConsequence
+ *
+ * @author Mr_Kywar mr_kywar@gmail.com
+ */
+class LimitlessStudieConsequence extends StudiesConsequence {
+    
+
+    public function execute() {
+        throw new ConsequenceException("Consequence-LSC : Not Yet implemented");
+    }
+
+}
