+++ modules/php/Game/Card/Consequence/Category/Studies/StudieLevelIncriseConsequence.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace SmileLife\Card\Consequence\Category\Studies;
+
+use SmileLife\Card\Consequence\ConsequenceException;
+
+/**
+ * Description of StudieLevelIncriseConsequence
+ *
+ * @author Mr_Kywar mr_kywar@gmail.com
+ */
+class StudieLevelIncriseConsequence extends StudiesConsequence {
+
+    public function execute() {
+        throw new ConsequenceException("Consequence-SLIC : Not Yet implemented");
+    }
+
+}
