+++ modules/php/Game/Card/Criterion/Factory/Category/Attack/BurnOutCriterionFactory.php
@@ -3,8 +3,8 @@
 namespace SmileLife\Card\Criterion\Factory\Category\Attack;
 
 use SmileLife\Card\Card;
-use SmileLife\Card\Consequence\AttackDestinationConsequence;
-use SmileLife\Card\Consequence\TurnPassConsequence;
+use SmileLife\Card\Consequence\Category\Attack\AttackDestinationConsequence;
+use SmileLife\Card\Consequence\Category\Attack\TurnPassConsequence;
 use SmileLife\Card\Criterion\CriterionInterface;
 use SmileLife\Card\Criterion\Factory\CardCriterionFactory;
 use SmileLife\Card\Criterion\JobCriterion\HaveJobCriterion;
