+++ modules/php/Game/Card/Criterion/Factory/Category/Attack/DismissalCriterionFactory.php
@@ -4,8 +4,8 @@ namespace SmileLife\Card\Criterion\Factory\Category\Attack;
 
 use SmileLife\Card\Card;
 use SmileLife\Card\Category\Job\Official\Official;
-use SmileLife\Card\Consequence\AttackDestinationConsequence;
-use SmileLife\Card\Consequence\DiscardConsequence;
+use SmileLife\Card\Consequence\Category\Attack\AttackDestinationConsequence;
+use SmileLife\Card\Consequence\Category\Generic\DiscardConsequence;
 use SmileLife\Card\Criterion\CriterionInterface;
 use SmileLife\Card\Criterion\Factory\CardCriterionFactory;
 use SmileLife\Card\Criterion\GenericCriterion\CriterionGroup;
