+++ modules/php/Game/Card/Criterion/Factory/Category/Studies/StudiesCriterionFactory.php
@@ -3,7 +3,8 @@
 namespace SmileLife\Card\Criterion\Factory\Category\Studies;
 
 use SmileLife\Card\Card;
-use SmileLife\Card\Consequence\LimitlessStudieConsequence;
+use SmileLife\Card\Consequence\Category\Studies\LimitlessStudieConsequence;
+use SmileLife\Card\Consequence\Category\Studies\StudieLevelIncriseConsequence;
 use SmileLife\Card\Criterion\CriterionInterface;
 use SmileLife\Card\Criterion\Factory\CardCriterionFactory;
 use SmileLife\Card\Criterion\GenericCriterion\CriterionGroup;
@@ -54,7 +55,7 @@ class StudiesCriterionFactory extends CardCriterionFactory {
         $studieLevelCriterion = new StudiesLevelCriterion($table, $card);
         $studieLevelCriterion->setErrorMessage(clienttranslate('You have already reached level ${level} of studies and you cannot exceed 6', ['level' => $actualLevel]));
 
-        return new CriterionGroup([
+        $criteria = new CriterionGroup([
                 //-- Limitless Studie
                 $limitlessCriterion,
                 //-- Classic Criterion
@@ -63,6 +64,10 @@ class StudiesCriterionFactory extends CardCriterionFactory {
                     $studieLevelCriterion
                 ], CriterionGroup::AND_OPERATOR)
             ], CriterionGroup::OR_OPERATOR);
+        
+        $criteria->addConsequence(new StudieLevelIncriseConsequence($card));
+        
+        return $criteria;
         
     }
